System solving
Unrelated but I tried using Antigravity (a coding assistant from Google). It was not fun. From now on there will be minimal LLM coding assistance as I did not find this session to be more productive that the rest, quite the contrary. Plus, getting to write the code line by line helps me better understand it.